Seperated helper classes from stripplot and fixed bug in plotpicker
